IngresoProducto parte  1

+++ src/main/java/com/siapa/dao/IngresoProductoDao.java
@@ -15,6 +15,6 @@ import org.springframework.stereotype.Repository;
  * @author Joao
  */
 @Repository
-public class IngresoProductoDao extends GenericDao<IngresoProducto, Long> {
+public class IngresoProductoDao extends GenericDao<IngresoProducto, Integer> {
     
 }
